feat: implement the authentication logic
