+++ src/auth/auth.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { AuthService } from './auth.service'
-import { LocalStrategy } from './local.strategy'
+import { LocalStrategy } from './strategies/local.strategy'
 import { UsersModule } from '../users/users.module'
 import { PassportModule } from '@nestjs/passport'
 import { JwtModule } from '@nestjs/jwt'
-import { JwtStrategy } from './jwt.strategy'
+import { JwtStrategy } from './strategies/jwt.strategy'
 import { AuthController } from './auth.controller'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { User } from 'src/users/entities/user.entity'
@@ -19,7 +19,7 @@ import { EmailModule } from 'src/email/email.module'
     PassportModule,
     EmailModule,
     JwtModule.register({
-      secret: process.env.JWT_SECRET || 'defaulySecretKey', 
+      secret: process.env.JWT_SECRET || 'defaulySecretKey',
       signOptions: { expiresIn: '60m' },
     }),
   ],
